Started implementing transactions

+++ Account.cpp
@@ -1,6 +1,6 @@
 #include "Account.h"
 
-Account::Account(const QString & name, int balance) :
+Account::Account(const QString & name, double balance) :
     m_name(name),
     m_balance(balance)
 {
