+++ Account.h
@@ -6,21 +6,21 @@
 class Account
 {
 public:
-    Account(const QString & name, int balance);
+    Account(const QString & name, double balance);
 
     QString name() const
     { return m_name; }
     void set_name(const QString &name)
     { m_name = name; }
 
-    int balance() const
+    double balance() const
     { return m_balance; }
-    void set_balance(int balance)
+    void set_balance(double balance)
     { m_balance = balance; }
 
 private:
     QString m_name;
-    int m_balance;
+    double m_balance;
     //TODO: Ccy
 };
